+++ src/components/Skeleton/GoogleAnalyticsProvier.tsx
@@ -3,7 +3,7 @@ import { useLocation } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import loginInfoDetailAtom from "atoms/loginInfoDetail";
 import reactGA from "react-ga4";
-import { gaTrackingId, nodeEnv } from "serverconf";
+import { gaTrackingId, nodeEnv } from "loadenv";
 
 const GoogleAnalyticsProvier = () => {
   const gaInitialized = useRef(false);
